refactor: remove TODO's
